refactor(cli): extract pure output helpers to src/cli/output.ts
Moves 8 self-contained stdout/formatting helpers out of cli/bakin.ts into a new
src/cli/output.ts (print, printTable, invocationCommand, normalizeUsage, usageLine,
statusIcon, formatBytes, daysUntil); cli/bakin.ts imports them back. No behavior change.

Deliberately scoped: the exit helpers (exitCommandIssue/Usage/UnknownSubcommand/
CommandFailure) stay in bakin.ts — they render through the printCommandIssueTui/
printCommandFailureTui wrappers, and printCommandFailureTui is also used by three non-exit
command sites, so the wrappers can't move yet. The emit() dispatcher, isTTY-branch rework,
and confirm-prompt unification remain out of scope (later focused step).

Verified: typecheck + lint clean; tests/cli 252/0 (--isolate); full binary build.
